Move StatementProtocolConfig out of network_service into json_rpc_service
