+++ light-base/src/json_rpc_service.rs
@@ -55,6 +55,8 @@ use alloc::{
 use core::{num::NonZero, pin::Pin};
 use futures_lite::StreamExt as _;
 
+pub use statement::StatementProtocolConfig;
+
 /// Configuration for [`service()`].
 pub struct Config<TPlat: PlatformRef> {
     /// Access to the platform's capabilities.
@@ -121,11 +123,7 @@ pub struct Config<TPlat: PlatformRef> {
     pub genesis_block_hash: [u8; 32],
 
     /// Statement protocol configuration. `None` if the statement protocol is disabled.
-    pub statement_protocol_config: Option<network_service::StatementProtocolConfig>,
-
-    /// Maximum number of seen statement hashes tracked per subscription for dedup.
-    /// `None` if the statement protocol is disabled.
-    pub max_seen_statements: Option<NonZero<usize>>,
+    pub statement_protocol_config: Option<StatementProtocolConfig>,
 }
 
 /// Creates a new JSON-RPC service with the given configuration.
@@ -166,7 +164,6 @@ pub fn service<TPlat: PlatformRef>(config: Config<TPlat>) -> Frontend<TPlat> {
                 system_version: config.system_version,
                 genesis_block_hash: config.genesis_block_hash,
                 statement_protocol_config: config.statement_protocol_config,
-                max_seen_statements: config.max_seen_statements,
             },
             requests_rx,
             responses_tx,
